pyproject.toml
[project]
name = "diff-tissue"
version = "1.1.0"
description = "Optimization of growing tissues using differentiable programming"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flax>=0.12.2",
    "jax[cpu]>=0.5.2",
    "jaxlib>=0.5.1",
    "jaxopt>=0.8.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "optax>=0.2.4",
    "optuna>=4.6.0",
    "pandas>=2.2.3",
    "pyyaml>=6.0.3",
    "scipy>=1.15.2",
    "shapely>=2.1.1",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.per-file-ignores]
"jax_bootstrap.py" = ["E402"]

[tool.mypy]
python_version = 3.12
files = ["src", "scripts"]
mypy_path = "src"
explicit_package_bases = true
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "9.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
    "types-pyyaml>=6.0.12.20250915",
]
